Refactor Polymarket into a lower menu and rename menu folders. fixes #32. (#40)
