+++ open_eth_terminal/CryptoMenu/actions/spot.ts
