+++ open_eth_terminal/CryptoMenu/model/CoinGeckoApi.ts
