+++ open_eth_terminal/CryptoMenu/model/index.ts
