+++ open_eth_terminal/CryptoMenu/types.ts
