+++ open_eth_terminal/PredictionMarketMenu/PolymarketMenu/model/Polymarket.ts
