+++ open_eth_terminal/PredictionMarketMenu/PolymarketMenu/model/index.ts
