+++ open_eth_terminal/StocksMenu/actions/alphavantage.ts
