+++ open_eth_terminal/StocksMenu/index.ts
