+++ open_eth_terminal/StocksMenu/model/index.ts
