+++ open_eth_terminal/StocksMenu/types.ts
